feat: добавить функционал для хранения и получения пользователей из БД

+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/InMemoryUserStorage.java
@@ -90,7 +90,7 @@ public class InMemoryUserStorage implements UserStorage {
   }
 
   @Override
-  public List<User> getFriends(int id) {
+  public List<User> getFriends(long id) {
     return getFriendsInfo(getUserById(id).getFriends());
   }
